add searchPage: refactor
