+++ src/index.js
@@ -3,7 +3,7 @@ import './index.scss';
 
 import { router } from './modules/router';
 
-import { mainPage } from './modules/mainPage';
+import { mainPage } from './modules/pages/mainPage';
 import { renderFooter } from './modules/render/renderFooter';
 import { renderHeader } from './modules/render/renderHeader';
 import { getData } from './modules/getData';
@@ -11,19 +11,19 @@ import { getData } from './modules/getData';
 import { API_URL, DATA } from './modules/const';
 import { createScssColors } from './modules/createScssColors';
 import { createElement } from './modules/createElement';
-import { categoryPage } from './modules/categoryPage';
-import { searchPage } from './modules/searchPage';
+import { categoryPage } from './modules/pages/categoryPage';
+import { searchPage } from './modules/pages/searchPage';
 
 const init = async () => {
   try {
+    DATA.navigation = await getData(`${API_URL}/api/categories`);
+    DATA.colors = await getData(`${API_URL}/api/colors`);
+    createScssColors(DATA.colors);
+
     router.on('*', () => {
       renderHeader();
       renderFooter();
     });
-
-    DATA.navigation = await getData(`${API_URL}/api/categories`);
-    DATA.colors = await getData(`${API_URL}/api/colors`);
-    createScssColors(DATA.colors);
 
     router.on('/', () => {
       mainPage();
