+++ src/modules/pages/mainPage.js
@@ -1,0 +1,11 @@
+import { renderNavigation } from '../render/renderNavigation';
+import { renderHero } from '../render/renderHero';
+import { renderGoods } from '../render/renderGoods';
+
+export const mainPage = (gender = 'women') => {
+  renderNavigation(gender);
+  renderHero(gender);
+  renderGoods('Новинки', {
+    gender,
+  });
+};
