+++ src/modules/pages/searchPage.js
@@ -1,6 +1,6 @@
-import { renderGoods } from './render/renderGoods';
-import { renderHero } from './render/renderHero';
-import { renderNavigation } from './render/renderNavigation';
+import { renderGoods } from '../render/renderGoods';
+import { renderHero } from '../render/renderHero';
+import { renderNavigation } from '../render/renderNavigation';
 
 export const searchPage = (data) => {
   console.log('data: ', data);
