+++ src/modules/render/renderHero.js
@@ -2,13 +2,13 @@ import { createElement } from '../createElement';
 import { TITLE } from '../const';
 
 export const renderHero = (gender) => {
-  console.log('HEROgender: ', gender);
   const hero = document.querySelector('.hero');
 
   if (!gender) {
     hero.style.display = 'none';
     return;
   }
+  hero.style.display = 'block';
 
   hero.className = `hero hero_${gender}`;
   hero.textContent = '';
