+++ src/modules/render/renderNavigation.js
@@ -10,8 +10,6 @@ let prevGender = '';
 export const renderNavigation = (gender, category) => {
   const navigation = document.querySelector('.navigation');
 
-  console.log(gender, category);
-
   if (!gender) {
     navigation.style.display = 'none';
   } else {
